requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==24.2.0
Authlib==1.3.1
bandit==1.7.9
billiard==4.2.0
black==24.8.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.6.1
cryptography==43.0.0
Django==5.1
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dparse==0.6.4b0
drf-spectacular==0.27.2
Faker==27.0.0
filelock==3.12.4
flake8==7.1.1
flower==2.0.1
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.10.0
idna==3.7
inflection==0.5.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.4.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pbr==6.0.0
pillow==10.4.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg==3.2.1
psycopg-binary==3.2.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.4
python-ulid==1.1.0
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
redis-om==0.3.2
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safety==3.2.5
safety-schemas==0.0.3
setuptools==72.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
stevedore==5.2.0
stripe==10.12.0
tornado==6.4.1
typer==0.12.3
types-cffi==1.16.0.20240331
types-pyOpenSSL==24.1.0.20240722
types-redis==4.6.0.20240806
types-setuptools==71.1.0.20240813
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
